fix: fix E2E test bugs and extract shared helpers
- Fix broken email construction in CR-1 test by storing email in variable
- Remove dead toggleBtn variable in SU-6 test
- Extract shared helpers (wait, go, register, login, dismissCookieConsent,
  createTestRunner) into new e2e/helpers.js module
- Refactor both test files to import from helpers.js instead of duplicating
